Itembox in /explore and /explore/[categoryName] now link to /explore/[categoryName]/postName
